Implemented Authentication Service
