+++ src/controllers/Authentication/generateToken.js
@@ -1,6 +1,6 @@
 const jwt = require("jsonwebtoken");
 
-const generateToken(content, secret){
+const generateToken = (content, secret) => {
     const jwtoken = jwt.sign(content, secret, {expiresIn : "1h"});
     return jwtoken;
 }
